requirements.txt

Flask==1.1.1
gunicorn==20.0.4
pymessenger==0.0.7.0
python-decouple==3.3